Refactor + add "add liquidity" quick action
