+++ packages/web/modals/index.tsx
@@ -1,6 +1,8 @@
+export * from "./add-liquidity";
 export * from "./base";
 export * from "./bridge-transfer";
 export * from "./connect-non-ibc-wallet";
+export * from "./create-pool";
 export * from "./ibc-transfer";
 export * from "./keplr-connection-selection";
 export * from "./lock-tokens";
