+++ packages/web/styles/globals.css
@@ -208,7 +208,7 @@
 
 /* Toasts override styles in `react-toastify/dist/ReactToastify.css` */
 .Toastify__toast-container {
-  z-index: 99 !important;
+  z-index: 200 !important;
   max-width: 360px !important;
   min-width: 360px !important;
 }
